+++ billingsoftware/src/main/java/learn/java/billingsoftware/service/impl/ItemServiceImpl.java
@@ -33,13 +33,13 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemResponse add(ItemRequest request, MultipartFile file) throws IOException {
-//        String imgUrl = fileUploadService.uploadFile(file);
-        String fileName = UUID.randomUUID().toString()+"."+ StringUtils.getFilenameExtension(file.getOriginalFilename());
-        Path uploadPath = Paths.get("uploads").toAbsolutePath().normalize();
-        Files.createDirectories(uploadPath);
-        Path targetLocation = uploadPath.resolve(fileName);
-        Files.copy(file.getInputStream(), targetLocation, StandardCopyOption.REPLACE_EXISTING);
-        String imgUrl = "http://localhost:8080/api/v1.0/uploads/"+fileName;
+       String imgUrl = fileUploadService.uploadFile(file);
+        // String fileName = UUID.randomUUID().toString()+"."+ StringUtils.getFilenameExtension(file.getOriginalFilename());
+        // Path uploadPath = Paths.get("uploads").toAbsolutePath().normalize();
+        // Files.createDirectories(uploadPath);
+        // Path targetLocation = uploadPath.resolve(fileName);
+        // Files.copy(file.getInputStream(), targetLocation, StandardCopyOption.REPLACE_EXISTING);
+        // String imgUrl = "http://54.254.162.86:8080/api/v1.0/uploads/"+fileName;
 
         ItemEntity newItem = convertToEntity(request);
         CategoryEntity existingCategory = categoryRepository.findByCategoryId(request.getCategoryId())
@@ -85,23 +85,23 @@ public class ItemServiceImpl implements ItemService {
     public void deleteItem(String id) {
         ItemEntity existingItem = itemRepository.findByItemId(id)
                 .orElseThrow(()-> new RuntimeException("Item not found: "+id));
-//        boolean isFileDelete = fileUploadService.deleteFile(existingItem.getImgUrl());
-        String imgUrl = existingItem.getImgUrl();
-        String filename = imgUrl.substring(imgUrl.lastIndexOf("/")+1);
-        Path uploadPath = Paths.get("uploads").toAbsolutePath().normalize();
-        Path filePath = uploadPath.resolve(filename);
-        try {
-            Files.deleteIfExists(filePath);
-            itemRepository.delete(existingItem);
-        } catch (IOException e) {
-            e.printStackTrace();
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Item not found");
-        }
-//        if(isFileDelete){
-//            itemRepository.delete(existingItem);
-//        } else {
-//            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to delete image item");
-//        }
+       boolean isFileDelete = fileUploadService.deleteFile(existingItem.getImgUrl());
+        // String imgUrl = existingItem.getImgUrl();
+        // String filename = imgUrl.substring(imgUrl.lastIndexOf("/")+1);
+        // Path uploadPath = Paths.get("uploads").toAbsolutePath().normalize();
+        // Path filePath = uploadPath.resolve(filename);
+        // try {
+        //     Files.deleteIfExists(filePath);
+        //     itemRepository.delete(existingItem);
+        // } catch (IOException e) {
+        //     e.printStackTrace();
+        //     throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Item not found");
+        // }
+       if(isFileDelete){
+           itemRepository.delete(existingItem);
+       } else {
+           throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to delete image item");
+       }
     }
 
 }
